update erlang.mk, remove checked in relx
Existing version of relx can not build working releases on OTP 21.
Erlang.mk has updated its internal relx since, but we are using a very
very old version of erlang.mk.

This commit updates erlang.mk, and removes the checked in relx.
Erlang.mk will download relx and use it when necessary. It is not
placed in the tree anymore.