Build with cuttlefish
We used to commit the cuttlefish binary directly into the git, but this way we have to update the binary every time the cuttlefish has code changes.

In this commit I removed the cuttlefish binary from the repo. When
buiding the project, the new built binary will be copied to the release directory.